# PROYECTO/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Hastechno",
        "description": "API para la gestión de productos, pedidos, usuarios, carritos y más para la tienda virtual Hastechno.",
        "contact": {
            "email": "babeltranp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Servidor de desarrollo local de la API"
        }
    ],
    "paths": {
...
        "/api/p/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener todos los pedidos del cliente autenticado",
                "operationId": "4f92637c24d234580f5fff353ab74087",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PedidoResponseDetailed"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Cliente no autenticado o sin información de cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no autenticado o sin información de cliente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Realizar un nuevo pedido",
                "operationId": "e3345cb1e28ab06bd61f79dcd767aa6a",
                "requestBody": {
                    "description": "Datos para crear un nuevo pedido.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido realizado y factura generada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido realizado y factura generada exitosamente."
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/PedidoResponseDetailed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores en la solicitud (producto no encontrado, stock insuficiente, cliente no asociado).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuficiente para: Producto X"
                                        },
                                        "errors": {
                                            "description": "Detalles de errores de validación si aplica.",
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Debes iniciar sesión para realizar un pedido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación de los datos de entrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor al procesar el pedido.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al procesar el pedido: Mensaje de error interno."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/p/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener un pedido específico del cliente autenticado por ID",
                "operationId": "dcfcc8ae7a87375c5075c4d746f97f42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido a obtener.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseDetailed"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Cliente no autenticado o sin información de cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no autenticado o sin información de cliente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado o no pertenece a este cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido no encontrado o no pertenece a este cliente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/p/mis-productos-comprados": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener todos los ítems de productos comprados por el cliente autenticado",
                "operationId": "9a5d6d0182c0d85317ef49ea8aafea6b",
                "responses": {
                    "200": {
                        "description": "Lista de ítems de productos comprados obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PedidoProductoWithProduct"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Cliente no autenticado o sin información de cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no autenticado o sin información de cliente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/pedidos/": {
            "get": {
                "tags": [
                    "Pedidos",
                    "Administración"
                ],
                "summary": "Obtener todos los pedidos con detalles para administración (solo SuperAdmin)",
                "operationId": "c11f1cc04fb9f18ed388320eadd03466",
                "responses": {
                    "200": {
                        "description": "Pedidos obtenidos exitosamente para administración.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PedidoResponseDetailed"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado (solo SuperAdmin).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PedidoResponseDetailed": {
                "title": "Pedido Detallado",
                "description": "Representación completa de un pedido con sus productos y factura.",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pedido"
                    },
                    {
                        "properties": {
                            "productos": {
                                "description": "Lista de ítems de productos en el pedido, con detalles del producto.",
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PedidoProducto"
                                        },
                                        {
                                            "properties": {
                                                "producto": {
                                                    "$ref": "#/components/schemas/Producto"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            },
                            "factura": {
                                "$ref": "#/components/schemas/Factura"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "PedidoProductoWithProduct": {
                "title": "Pedido Producto con Detalles de Producto",
                "description": "Representa un ítem de producto dentro de un pedido, incluyendo los detalles completos del producto.",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PedidoProducto"
                    },
                    {
                        "properties": {
                            "producto": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT con el prefijo 'Bearer ' (ej: 'Bearer eyJ0eXAiOiJKV1Q...') para acceder a los endpoints protegidos.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}